debug et test validePaiement

+++ src/main/java/com/analytique/gestion_analytique/Repositories/PayeDetailsRepository.java
@@ -4,7 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.analytique.gestion_analytique.Models.PayeDetails;
 
-public interface PayeDetailsRepository extends JpaRepository<PayeDetails, Long> {
+public interface PayeDetailsRepository extends JpaRepository<PayeDetails, Integer> {
 
     
 }
